Move few methods to proper place
